+++ R/arrows3d.R
@@ -43,8 +43,7 @@
   c(a[2]*b[3]-a[3]*b[2], -a[1]*b[3]+a[3]*b[1], a[1]*b[2]-a[2]*b[1])
 }
 
-# draw a cone (e.g. tip of an arrow)
-#' Title Draw a 3d cone
+#' Title Draw a 3D cone
 #'
 #' @param base   coordinates of base of the cone
 #' @param tip    coordinates of tip of the cone
@@ -54,7 +53,7 @@
 #' @param ...    rgl arguments passed down
 #'
 #' @return       returns the integer object ID of the shape that was added to the scene
-#' @author       January Weiner, from pca3d
+#' @author       January Weiner, borrowed from from \code{\link[pca3d]{pca3d}}
 #' @export
 #' @import rgl
 #'
@@ -85,11 +84,16 @@ cone3d <- function( base, tip, radius= 10, col= "grey", scale= NULL, ... ) {
 }
 
 
-#' Draw 3d Arrows
+#' Draw 3D arrows
 #'
-#' Draws nice 3d arrows with \code{cone3d}s at their tips
+#' Draws nice 3D arrows with \code{cone3d}s at their tips
 #'
-#' @param coords     A 2n x 3 matrix giving the start and end (x,y,z) coordinates of n arrows, in pairs
+#' This function is meant to be analogous to \code{\link[graphics]{arrows}}, but for 3D plots using \code{\link[rgl]{rgl}}.
+#' \code{headlength}, \code{scale} and \code{radius} set the length, scale factor and base radius of the arrow head, a
+#' 3D cone. The units of these are all in terms of the ranges of the current rgl 3D scene.
+#'
+#' @param coords     A 2n x 3 matrix giving the start and end (x,y,z) coordinates of n arrows, in pairs.  The first vector
+#'                   in each pair is taken as the starting coordinates of the arrow, the second as the end coordinates.
 #' @param headlength Length of the arrow heads, in device units
 #' @param head       Position of the arrow head, e.g., "end"
 #' @param scale      scale factor for base and tip of arrow head
@@ -97,7 +101,8 @@ cone3d <- function( base, tip, radius= 10, col= "grey", scale= NULL, ... ) {
 #' @param ...        rgl arguments passed down to \code{segments3d} and \code{cone3d}
 #'
 #' @return           none
-#' @author       January Weiner, from pca3d
+#' @author           January Weiner, borrowed from from \code{\link[pca3d]{pca3d}}
+#' @seealso          \code{\link{vectors3d}}
 #' @export
 #'
 #' @examples