+++ NEWS.md
@@ -4,6 +4,7 @@
 - added `vec()` function
 - added `is_square_matrix()` tests
 - added `power_method()`, power method for dominant eigenvector (thx: Gaston Sanchez)
+- added `arrows3d()` for 3D geometric diagrams
 
 
 # matlib 0.5.2